Directly convert frame name tails to Int32
This is a little optimization for large mail merge jobs with many
frames. The current alorithm doesn't only try to produce unique but
also reasonable names for the frames.

Per se the algorithm is horrible ineffective for large numbers of
frames, but this is impossible to fix without changing the
underlying vector of frames to a sorted one to find and check just
the correctly prefixed frames.

This patch directly converts the frame names tail instead of
creating substring of the number tail, which saves billions of
malloc and free calls seen with my sample document.

Change-Id: Iefdee4053480f40f106c49867bc5a64ec207ba1b