Optionally paste the full document
In the mail merge case we want to paste the full document including
the SwTxtNode containing the SwFmtPageDesc node, which is always
the first node of a document.

For normal paste we want to merge the first paragraph into the
paragraph at the cursor position. This currently loses all Flys
anchored at the first paragraph.

Change-Id: I2b1ce279743f559e700b47276012ad56d74a059c