MM: fix lost or broken paragraph anchored objects
With the extended SwFEShell::Paste version, we can finally paste
full documents without losing the paragraph anchored Flys of the
first paragraph.

Actually one should be able to drop the InsertPageBreak and the
single paragraph delete, but my attempt resulted in problems with
page anchored Flys and somehow reintroduced the problem seen in
fdo#80395 with draw objects.

Change-Id: I370fa89f432ba4dd95c1f02605f33f04dfbaad85